pyproject.toml
[project]
name = "lyricflow"
version = "0.1.0"
description = "A modular Python toolkit for automated processing and embedding of song lyrics"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "LyricFlow Contributors"}
]
license = {text = "MIT"}
keywords = ["lyrics", "romanization", "japanese", "music", "audio", "metadata"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fugashi[unidic-lite]>=1.5.1",
    "pykakasi>=2.3.0",
    "mutagen>=1.47.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
]
openai = [
    "openai>=1.10.0",
]
gemini = [
    "google-generativeai>=0.3.0",
]
whisper = [
    "openai-whisper>=20231117",
    "torch>=2.0.0",
]
fetch = [
    "requests>=2.31.0",
]
tui = [
    "textual>=0.50.0",
    "requests>=2.31.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
all = [
    "lyricflow[api,openai,gemini,whisper,fetch,tui]",
]

[project.scripts]
lyricflow = "lyricflow.cli.main:cli"

[project.urls]
Homepage = "https://github.com/yourusername/lyricflow"
Repository = "https://github.com/yourusername/lyricflow"
Issues = "https://github.com/yourusername/lyricflow/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lyricflow"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

